add: exception_parser_test.dart
